Replaces updateLabel with updateField
This allows the entire field to be updated instead of just the label,
paving the way for more props (ie. placeholder, errors, options, etc).

I also introduced internal Ids to the fields as a constant way to
reference them instead of relying on indexes or other things (ie. field
name) which will be prone to change.